(PUP-4116)(PUP-4117) Add json-schema and description of catalog diff
This adds the json-schema for a catalog delta and a description
of its semantics. It also adds the json-meta-schema itself (to
make it available for tests and users).

The catalog.json is the catalog schema (a copy of the schema in
puppet), and error.json a schema for errors. These two may be removed
to instead rely on puppet containing them. It is uncertain what should
happen at the next major puppet release / if there are changes to the
catalog as puppet is likely to change its copy to refelect the latest
catalog - the preview may need to support multiple versions.