(PUP-4149) Regenerate module skeleton
Prior to this commit, the module skeleton had been generated with
the puppet-module-skeleton tool but with a lot of the content
removed. Something was missing that was causing an error when
`puppet module list` was called with preview in the modulepath.
Rather than try to track down exactly what was causing the issue,
just regenerate the module skeleton and leave the structure mostly
intact.