(PUP-4128) Add migration check for "significant whitespace"
This adds a migration check for "significant whitespace" (in
a round about way) since it warns about the effect of whitespace
when that causes an Array to appear as a resulting value last in
a block. Other cases are caught as non-productive expression errors
and do not need a migration check.