(PUP-4119) Make puppet --preview node output migration issues
This enables a migration checker around a compilation and dumps
the resulting warnings to stdout.

This is not the final solution but shows that it works to
pass the migration checker via the context through to the
compilation. (In the real solution this is done in the new
dual compile indirection, which is not yet written).