(PUP-4118) Add indirection that compiles twice for baseline and preview
This adds an indirection that compiles two catalogs; one for
the baseline (the environment that is configured for the node
using the same mechanism to determine the node's environment
as 'puppet master --compile'), and one for the preview environment
given on the command line.

The output is written to a ./PREVIEW_OUTPUT directory from where
the command is given. This is a temporary solution as not all
pieces are in place to write to the expected location
as defined by a Puppet setting.