(maint) Add edge counts to delta, and compliant to conflicting resource
This adds summary edge counts for preview and baseline catalogs to
make it easier to write json queries/filter out without having to
count.

This also adds a compliant flag to conflicting resources that
indicates that although there are differences in the resource's
attributes, they are all compliant. This saves always having
to scan all conflicting attributes to see if a conflicting
resource is compliant.