(PUP-4280) Link options to arguments to catalog diff
This passes the argument --skip_args to the catalog diff.

The catalog diff supports a verbose mode where attributes
of missing and added attributes are also included in the diff.
There was no option for this on the command line. This commit adds
the option --verbose_diff (-vd) and passes the option on to the
catalog diff producer.