(maint) Fix problem when switching environment to future
If both environments (baseline and preview) are future parser
environments there was no problem. When switching from
current parser to future parser the check if future parser is
activated in Puppet failed becuse it checks the parser for
the :current_environment, and it was not overridden.

This changes the order of overriding and switching the
node's environment and passes the resulting environment as
:current_environment.