(PRE-17) Handle trusted node data from fact terminus
Without this a request to preview would always have
trusted node information as if the node was local (no extensions
parsed from a cert).

This is problematic since both node classification and compilation
may depend on trusted data; and preview would then compile two
catalogs that may have different content than what agent sees.

This fix allows trusted node data from the fact terminus to ressurect
the trusted node data as it was originally presented in an agent
request. The "authenticated" key is however set to false unless the
user running the preview is root, and has set the option --trusted.

A root user should only do this if the user trusts the puppet
configuration and the terminus used for storing and retreiving
facts. The root user is trusted to make this choice.

Users that make use of trusted node information should check the
authenticated flag in their manifests when creating a catalog
containing sensitive information that was entered into the catalog
in clear text based on information in the trusted information. These
users will need to verify preview/migration of such sensitive catalogs
by letting a user with root privileges run the preview.

Users that do not have such sensitive infomation in their catalogs, but
do depend on trusted information can use the command both as a regular
user or as root, and do not have to set --trusted (unless their
manifests check the authenticated value and produce different catalogs
(or errors)).