(PRE-19) Add ability to suppress number/string differences
The 3x parser uses strings as the internal representation of numbers.
This will potentially create a lot of noice when creating deltas
between 3x and 4x compiled code. This commit adds the boolean option
--ignore_string_numeric_diff. When used, the CatalogDelta producer will
detect and ignore differences between numbers in the preview model
that are equal to their string counterbart in the baseline if the
string is converted to a numeric.