(maint) Fix problem with errors being swallowed
The problem was that errors were logged but immediately followed
by an exit.

Strange things happen when a method is directly using "resuce" and
"ensure" and those are not for a begin/end block but for the method
itself.

Now, there is a begin/end block and handling of general exception
that occur outside of the compilation of baseline and preview. (Those
where the errors that were swallowed).