(maint) Change UX of ignoring str-numeric diff
Before this, the UX was to use the option `--ignore_string_numeric_diff`
to get rid of many false conflicts. Since that is the default when
using `--migrate` it was just too much to type (and understand).

Now, this mode is on by default when using `--migrate`, and it
can intead be turned off with `--diff_string_numeric`, but only in
combination with `--migrate`.