(PRE-22) Ensure that attribute that are Sets can be single values
The attributes before, after, subscribe, notify, and tags use
Set semantics when comparing baseline and preview values. Before
this commit it was assumed that the values to be compared originated
from arrays. This is now changed so that single values are also
accepted (and will be converted into one element arrays) and compared
using Set semantics.