(PRE-14) Add spec test that validates the catalog diff with json schema
This commit adds the tests needed to validate the json schema for
the CatalogDelta. It also altered all current tests using CatalogDelta
instances so that the instance is validated using the schema.

The CatalogDelta was modified so that the method to_hash now produces
a Hash that, when JSON generated, is a valid instance of the catalog
delta json schema. This made the delta easier to validate and it
also vouches for cleaner code since the hash doesn't need to be
modified with additional attributes after the delta is created.

In fact, there's no real reason why the preview command should use
a hash everywhere instead of the proper CatalogDelta object. The
only time the hash is really needed is when the user requests JSON
output. This commit did not change this however, but I suggest that
change to become a separate issue.