(PUP-4241) add file checks to puppet preview integration tests
This change tests the various log files that are output from puppet
preview.  They should all, always be there, no matter the exit code, but
are tested against the successful run with exit code of 0.

This change also changes "default" to master, to be more definitive
and adds a debian configuration file

This change removes --migrate option from a single test to ensure it's
not required and fixes the test for upstream changes to stderr features.