(maint) update integration helper for pe
Changes to ensure integration setup works for both foss and PE.
Remove hard-coded path to moduledir, this is done correctly when
options[:type] is set to foss or the default of pe.
Add two example pe host config files to ease setup outside of CI
fix the foss host config file to specify type.

This won't work until the CI job is updated to specify a PE 3.8 build,
and SPEC_FORGE is unset (or the module is on the staging forge)