(PRE-23) Add a copy of the PuppetDB Facts terminus
This patch adds a copy of the 2.3.2 PuppetDB Facts terminus that contains one
change: the `find` method hits the v4 API instead of the v3 api. Using the v4
API allows Preview compile to fetch structured facts that are not stringified.
The use of non-stringified facts is important as future parser comparison
operators are type sensitive. This commonly manifests in comparisons between
stringified facts and integers which will fail when the future parser is in
use.