(PRE-26) Make --view xxx_log output pretty json
Before this the --view of a log produced raw json outout which was
difficult to read. Other --view options displayed pretty output because
the files are written in pretty format in the first place.

This is now fixed by loading the logfile as json and prettyfying it on
output.

This also contains a small maintenance fix of unused "pretty_json" flag
to the display_log method. This was unused as that method only displayes
the essuential information severity, message and location for output to
the console when compilation fails.