(PRE-27) Updating reporting based off PR feedback
Update where we report the results for a preview run. In order to
ensure we can still print results for a single node run, always store
the last catalog diff produced in an instance variable. This is a
simple solution and will be refactored later when we refactor the
way output works.