(PRE-33) Add a Comprehensive Overview model
This commit adds a new Overview model that is capable of storing
data from several CatalogDelta instances. The data is stored in a
highly normalized fashion to enable it to be queried in various
ways.

The Overview model is capable of serializing/deseralizing itself
using JSON although due to the nature of its content, the JSON is
hardly readable (it's mostly integer id refrences pointing to other
entries within the same structure).

The model can be queried and traversed using normal Ruby code. This
is accomplished by wrapping the actual entities in an object that
translates unrecognized method calls to logic that resolves
a relationship. A special WrappingArray is used to wrap collections
of entities and intermediate entities (many-to-many relationships)
are made invisible.

See spec/unit/overview_model_spec.rb for examples on how to write
queries.