(PRE-29) Make --migrate take an argument 'migration'
This adds an argument to the --migrate option. (The previously supported
migration kind is known as '3.8/4.0').
This commit also adds basic checking of the migration kind against
either Puppet::VERSION (since puppet 3.8.0 does not have a
MigrationChecker that can answer what it supports in terms of migration
checks). Later version has a method to return what kinds of checks the
runtime supports (with calls to do the reporting in Catalog Preview).

Code is added to support geting a decent error message when using
Catalog Preview with versions after latest on 3.x (i.e. the given
migration kind must both be known by Catalog Preview, (so it knows how
to handle it), and supported by the Puppet runtime's MigrationChecker.
User is directed to help if that is not the case after pointing out how
the given migration kind is wrong for the current configuration.