(maint) Remove beaker `step` usage
Prior to this we had a couple errorious usages of Beaker's `step` method
(a method for organizing tests conceptually similar to RSpec's usage of
`context`).

I don't believe using this method with beaker-rspec was
ever intended to be supported, though it works in *some* versions of
Beaker.

It does not work post Beaker 2.12.0 and this commit removes
usage of `step` so that it can support Beaker post 2.12.