(PRE-36) Add --view summary for multiple nodes
Add functionality so that if the user is running with --view summary
(which is the default) and multiple nodes, a short one line summary
of each node will be printed. Each summary is simply "status: node name"
so for example "PREVIEW FAILED: foo.local".

Failures will be printed in red, equal catalogs printed in green,
and all other statuses printed in the default color.