(PRE-33) Use CatalogDelta model instead of Hash
The current code creates a CatalogDelta model and then immediately
turns it into a hash. Subsequent code then uses this hash with
string constants to lookup all values. This is bad design. There's
no need to create this hash (unless serialising the full model into
JSON) and the hash is also difficult for an IDE to validate.

Another reason to not use the hash is that the Overview model can
merge CatalogDelta instances. It cannot merge hashes.

This commit changes the use of the Hash to instead use proper methods
on the CatalogDelta model.