(PRE-36) Add --view summary for multiple nodes
Prior to this commit, when you ran preview with --view summary for
multiple nodes it would only show a brief abstract. In order to
provide the user with more helpful output, add functionality to
provide a brief summary of the status of each node.

Additionally delete a few methods previously used for sorting which
are no longer needed due to the new data model.