(PRE-35) Made --last work with multiple nodes
Add functionality so that you can run --last and get data for
multiple nodes. You can either specify which nodes you would like
output for (one or several), or do not specify and receive the data
for all nodes which have data in the preview_outputdir. You still must
using a compatible --view flag with --last.