(PRE-47) Improve error handling around --last
Prior to this commit, there were some errors that could be encountered
when running --last which we were not catching. This meant the user
would see the Ruby error which did not include a helpful description.

Now when an incompatible --view option is used, or when data for the
specified node does not exist, preview will raise more helpful errors.