(PRE-55) Print summary even without catalog
Prior to this commit, nothing would be printed when running with
--view summary if no catalog delta was produced (e.g. if compilation
failed). Since this is not very helpful to the user, update the
display_summary method so it will print something even if there
is not catalog delta.

We can't generate the summary from @overview since all the
interesting information comes from the catalog delta, so just print
the last line of the summary.