(PRE-61) Ensure that environment is propagated on compile failure
Prior to this commit the environment was never stored in any file
unless the diff succeeded. A consequence of that was that the
overview report didn't display correctly.

This commit adds a new file named compile_info.json that stores
information about the compile (time, exit_code, and environment
names) regardless of the outcome of the compilation. This info
is then used when recreating the overview by using the --last
option.