(PRE-67) Update preview and puppet versions for 2.0.0
Prior to this commit, the metadata required puppet and PE 3.8.0.
Since 3.8.0 contains a bug that prevents you from setting the parser
to future in environment.conf, we don't want users to try and use that
version of puppet with preview (since it will prevent them from using
the suggested workflow for migration). Update the metadata so it requires
version 3.8.1 instead.

Also update the preview version to 2.0.0 for the new release.