(PRE-44) add multiple node tests
This change adds multiple nodes to most of the preview runs that make
sense in the existing tests. We specify nodes on the command line and
via a file using --nodes.
Please note the two TODO: comments. This behavior seems unexpected to
me.
Some of the code herein is quite repetitive, suggest ticketing to clean
up and/or dry-out in next commit. Could be cleaned up in this PR with
some effort, probably only useful if wanting to get this in prior to
sprint-end.