(PRE-70) Warn users when migrating with 3.8.0
Since there is a bug in puppet 3.8.0 which prevents users from
setting their parser per environment, users cannot use the desired
migration work flow with PE 3.8.0. In order to make sure they are
not seeing failures and don't understand why, add a warning if their
PE version is 3.8.0 and they are using --migrate.