Remove vagrant-wrapper
The vagrant-wrapper gem is not actually required for anything; if
installed, it monkey patches Object and causes bundler to behave in
strange ways, including saying that it is required when it is not. I
recommend removing it from all Gemfiles and uninstalling it from your
system.

We jokingly call it a virus of the rubygem world, spreading by means of
pull requests :) (https://github.com/bfraser/puppet-grafana/pull/16)