(PRE-76) Make preview use original pdb fact indirection for pdb 3
Before this, if the fact indirection was puppet_db a PDB v4 REST API
indirector was used to get correctly typed facts. Now that PDB 3 is
released and uses the v4 APIs by default we should use the original
indirection.

As a consequence of the change in the PDB fact indirection catalog
preview would fail since the modified copy of the previous indirector is
making calls to Puppet::Utils::Puppetdb and invoking methods that are no
longer present in PDB 3.

Since Catalog Prevew should work with both PDB 2 and 3, the change is to
not install the special fact indirection if PDB is 3 or newer.