(PRE-75) Add support for 4.x catalog schema
Before this commit, Catalog Preview would end up comparing two empty
sets as it compared the content of the catalog's 'data' element. In 4.x
(after PUP-3355) the extra data outside of the 'data' tag was removed
and the content of 'data' was lifted to its parent element.

This changes the logic for creating the diff by drilling down into
'data' if it exists in the catalog. This also means it is possible to
compare a 3.x and a 4.x catalog.

Unfortunately there was no new meta-data (such as schema version) added
to the schema and the only way to detect if this is a 3.x or 4.x format
is to check if there is a 'data' element or not.