(PRE-78) Make attributes with Set values be presented as lists
The Ruby JSON implementation in use by preview does not present a
ruby Set as a JSON list. Instead, the class name and instance id
of the Set is printed. This data is not useful in a catalog diff.

This commit ensures that all such Set instances are converted into
sorted Arrays once the delta calculation has been made, thus ensuring
that they will be output as JSON lists in when the delta is printed.