(maint) refactor gemfile
This change solidifies our use of bundler groups in the single Gemfile for
reduction of the bundled gem dependencies for dev,unit tests and
integration tests.  This should save time in the unit and integration
jobs.
This change also adds a user/global gemfile merge which allows for easy
addition of things like pry and other debugging gems from a central
location for all projects.  yay

[skip ci]