(PRE-66) Improve exception handling and exit codes
This commit splits the `#run_command` into two methods so that the
`#run_command` is responsible of catching a general exception and
do an `#exit`. The method that it calls is named `#main` (this is
consistent with other Puppet applications). The `#main` method will
not do an exit and it will not catch general exceptions.

A special exception was introduced to replace the practice of just doing
raise without an explicit exception in the `#main` method.

The main driving force behind these changes is to make it easier to
write unit tests.