(PRE-66) Make handling of node argument consistent
Before this commit, a single `node` was sometimes set in the `options`
map and later accessed in various places in the code. This lead to
problems because sometimes the code path didn't set the option before
it was used. One place where this could be observed was when printing
a summary using `--last`.

This commit ensures that the methods that require a `node` also has
a proper `node` parameter instead of relying on it being passed in
the `options` hash and that the parameter is propagated as needed.