(PRE-72) Use process owner as default for Puppet user and group
Puppet versions < 4.0 will make an attempt to change the user and group
of a log destination file to 'puppet'. This fails unless the user has
root privileges. This commit changes the catalog_preview command so that
it detects when the process owner is non-root user and then sets the
`Puppet[:user]` and `Puppet[:group]` to values that correspond to the
`uid` and `gid` of the process owner, thus preventing permission errors
further on.