(PRE-91) Add flag to suppress value vs [value] conflicts
This commit adds a new command line option `diff-array-value` which
defaults to `true` (to retain backward compatibility) but can be turned
off by using `--no-diff-array-value`. When it is turned off, a value
in the baseline catalog will be considered equal to a one element array
containing that value in the preview catalog.

The option can only be used together with --migrate 3.8/4.0.