(PRE-86) Deprecate CLI option --trusted
Since PUP-5061 was fixed, there's no longer a reason to have a --trusted
option. This commit removes the logic that was sensitive to this option
and adds a deprecation warning in case it is used. The warning is
printed conditionally to avoid that it corrupts valid JSON output.

The commit installs a copy of the Puppet::Parser::Compiler#sanitize_node 
method that is copied from a 4.4.0 version of Puppet.