(PRE-93) Change CLI options to use dash instead of underscore
All CLI options in the Puppet code-base uses dash as word separator.
Before this commit, catalog-preview used an underscore. This commit
changes this so that both dash and underscore can be used (to retain
backward compatibility). The documentation is changed to only mention
the dash variants.

The commit also changes boolean options so that they can be prefixed
with 'no-' to negate their meaning. This will be relevant when we add
boolean options that default to `true` (as the planned
--diff-array-value for instance) and we want to provide a way to turn
the option off.