(PRE-100) Fail compilation of nodes that doesn't have a valid factset
Prior to this commit, a node that didn't have a valid factset would be
compiled anyway and cause a lot of errors. This made the output hard
to read. This commit ensures that the compilation of such nodes fails
with immediately with a single error indicating that facts seems to
be missing.