(PRE-101) Add flag to control node list output
This commit adds the flag --[no-]report-all. If true, the human
readable overview report (obtained with --view overview) will include
a list of all nodes, as opposed to a list limited to the top ten nodes
with most issues.

The default setting for the flag is `false`, i.e. only the top ten nodes
will be reported.