(PRE-95) Add flag to skip/not skip nodes that are inactive
Prior to this commit, a compile was attempted for all nodes given on
the command line. Now, a check is instead made to see if the 'status'
action exists on the 'node' face. This action is provided by PuppetDB
and if it exists, it is used to obtain the status for each of the nodes.

Nodes that are found inactive will not be compiled unless the new option
--no-skip-inactive-nodes is given. Nodes that are not found at all will
not be compiled and result in an error printout. The preview command
will then continue with the next node.