(PRE-103) Improve sorting and information in "all nodes" list.
This commit replaces the simple "issue count" in the node list
with five counters for errors, warnings, added resources, removed
resources, and changed resources. It also changes the sorting of the
list to a descending sort based on errors, warnings, and the sum
of all resource counters.

The output produced by `--view overview` now outputs a table with node
name and counters for errors, warnings, and the sum of all resource
counters instead of just the node name and an issue counter.