(PRE-100) Fix failing test for 'No osfamily fact found'
The code assumed that the compile call would accept a block. That is
only true when using the call on a compiler instance, not on the class.
The subsequent test only "worked" on 4.x (since it was disabled there).

This commit ensures that the check for empty factset is in the correct
place and that the test works when using Puppet 3.8.x