(PRE-95) remove skip-inactive-nodes docs
The `--no-skip-inactive-nodes` flag is not yet functional and the
`--skip-inactive-nodes` behavior will be the default. So, the switch
currently does not change the behavior of catalog preview.

This commit removes the documentation for the switch from the README
and the `help` output. This documentation can be reinstated when the
`--no-skip-inactive-nodes` flag works.