(PRE-104) Fix acceptance baseline
This commit updates the `skip-inactive-nodes` acceptance tests.

Some of the tests were attempting to use the `--view baseline` flag
when providing a node file with multiple nodes. This has been changed
to use the `--view none` flag.

The `puppetdb-version` was not being set in this test block to
make sure that the properly configured API was being used. This value
is now being set based on the `puppet-version`.

The set up steps in the context did not seem to execute consistently
prior to the individual tests. These have been wrapped in an rspec
`before(:all)` block to ensure that the setup steps are completed
before the tests commence.