(PRE-122) Inactive nodes not skipped by default
This commit ensures that the method Preview#using_puppetdb? continues
to return `true` for the duration of a run. This was not the case
earlier since that method tested for the name of the Facts terminus and
when it was found to be 'puppetdb', that termnus was replaced with
the tailored 'diff_puppetdb' terminus. Subsequent calls to
test for if a node was active or deactivated.