(PE-11711) Exclude puppet_enterprise modules from preview
Prior to this commit, catalog preview would attempt to analyse
the modules that are included with Puppet Enterprise. Since
customers are not expected to fix problems in code that they
don't own, and since the diffs in these modules are known to
be "safe", differences in them should not clutter the output.

This commit excludes modules that begin with "puppet_enterprise"
or "pe_" from being included in the diff analysis.