(maint) Tie rubocop versions to appropriate ruby versions
Prior to this commit, the Gemfile installed the latest version of
rubocop on all ruby versions. Because the latest version of
rubocop requires ruby 2.0 or greater, this was causing failures in
CI when spec tests were run against ruby 1.9.3.

In order to prevent this issue, only install the latest version
for ruby 2.0 or greater and for older versions of ruby use a
compatible release of rubocop.