(maint) Pin listen gem version to prevent CI failures
Prior to this commit the listen gem (which the listen-guard gem
depends on) was not pinned to a particular version. Since the
latest version of the gem requires a Ruby version greater than 2.2
it was causing unit tests to fail on older versions.

Pin the listen gem to an older version so we can run with older
Ruby versions.