(maint) Update Travis CI Ruby versions
Prior to this commit preview was running in Travis CI with three
Ruby versions that are all no longer supported by Puppet. Update
the .travis.yml file to use more current versions. Soon this will
be module sync'd and this file will be managed that way, but in
the mean time this will help unblock work for preview.

Also remove the exclusion of development gems on the bundle install.
We need to install development in this case because that is where the
safe_yaml gem is pinned to 1.0.4. If we don't have this pin then we
will install the version of the gem vendored in Puppet and will
encounter a Syck dependency error (see PUP-3796).